changed the adios_define_var return type. Added adios_write_byid api.

+++ src/core/adios_internals.h
@@ -456,12 +456,12 @@ int adios_common_declare_group (int64_t * id, const char * name
                                ,enum ADIOS_FLAG stats
                                );
 
-int adios_common_define_var (int64_t group_id, const char * name
-                            ,const char * path, enum ADIOS_DATATYPES type
-                            ,const char * dimensions
-                            ,const char * global_dimensions
-                            ,const char * local_offsets
-                            );
+int64_t adios_common_define_var (int64_t group_id, const char * name
+                                ,const char * path, enum ADIOS_DATATYPES type
+                                ,const char * dimensions
+                                ,const char * global_dimensions
+                                ,const char * local_offsets
+                                );
 
 int adios_common_define_var_characteristcs  (struct adios_group_struct * g, const char * var_name
                             				,const char * bin_interval
